requirements.txt
requests
PyPDF2
pycryptodome
openai
python-dotenv